Split out isSitePlanDiscounted
